% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bin.fi.exact}
\alias{bin.fi.exact}
\title{Calculate an exact fragility index for the fisher exact test}
\usage{
bin.fi.exact(x, get.p, alpha = 0.05, max.f = Inf, verbose = FALSE)
}
\arguments{
\item{x}{a 2x2 matrix with treatments on rows and outcomes on columns}

\item{get.p}{a function inputting a table and outputting a p-value}

\item{alpha}{a string whose value being "exact" leads to an exact calculation}

\item{max.f}{the maximum fragility index considered}

\item{verbose}{a boolean value for whether to print progress
percentages (in increments of roughly ten percent)}
}
\value{
a fragility index
}
\description{
Exactly alcululate a fragility index using for the fisher exact test using a table.
Do a grid search over all FI values <= some value.
}
\details{
Do not call this function: use bin.fi with option 'exact' instead
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, rgeom(4, 1 / 50))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("control", "treatment")
get.p <- function(tab) fisher.test(tab)$p.value
bin.fi.exact(x, get.p)
}
