% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bin.fi.exact2}
\alias{bin.fi.exact2}
\title{Calculate an exact fragility index for the fisher exact test (2.0)}
\usage{
bin.fi.exact2(
  x,
  get.p,
  alpha = 0.05,
  fi.start = 1,
  can.vary = matrix(rep(TRUE, 4), nrow = 2),
  start.reversed = FALSE,
  start.p = NA
)
}
\arguments{
\item{x}{a 2x2 matrix with treatments on rows and outcomes on columns}

\item{get.p}{a function inputting a table and outputting a p-value}

\item{alpha}{a string whose value being "exact" leads to an exact calculation}

\item{fi.start}{The starting fi considered, by default 1.}

\item{can.vary}{a 2x2 boolean matrix for whether each entry can change.}

\item{start.reversed}{a boolean for whether the start.fi value already reverses
statistical significance. Note, FALSE just means unknown.}

\item{start.p}{The p value of the table with reversed statistical singificance, paired with
the parameter start.reversed}
}
\value{
a fragility index
}
\description{
Exactly alcululate a fragility index using for the fisher exact test using a table.
Start at a given FI value and radiate out from there.
}
\details{
Do not call this function: use bin.fi with option 'exact' instead. If bin.fi.exact2 is
initialized with too large of a value, the algorithm could malfunction. It's best to
leave the default behavior unless sure.
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, rgeom(4, 1 / 50))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("control", "treatment")
get.p <- function(tab) fisher.test(tab)$p.value
bin.fi.exact2(x, get.p)
}
