% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{bin.fi.incidence}
\alias{bin.fi.incidence}
\title{Loop over bin.fi for relevant grid of min.p values}
\usage{
bin.fi.incidence(
  data.table = NULL,
  X = NULL,
  Y = NULL,
  alg = "exact",
  test = "fisher",
  fi.start = NULL,
  alpha = 0.05,
  verbose = FALSE,
  warm.start = TRUE,
  delta = NULL
)
}
\arguments{
\item{data.table}{a 2x2 non negative integer matrix representing the output of a clinical trial.
The rows correspond to treatment/control and the columns to event/nonevent.}

\item{X}{a data frame of covariates}

\item{Y}{a data frame of responses}

\item{alg}{a string specifying the FI algorithm, 'greedy' (default), 'exact', 'original', or 'original.bothdir'}

\item{test}{a string specifying the test, defaulting to 'fisher' for the Fisher exact test.
An alternative is 'fisher.midp' and 'pearson.chisq'.}

\item{fi.start}{the starting fragility index considered in the exact algorithm.}

\item{alpha}{a number for the size of test}

\item{verbose}{a logical value for whether to print status updates while running}

\item{warm.start}{If using the exact algorithm without specifying a \code{max.f} argument,
a greedy algorithm will be ran first to upper bound the search for the exact algorithm.
Set to FALSE to only run the exact algorithm.}

\item{delta}{the noninferiority margin for when test='ni.normal'}
}
\value{
a matrix containing the p values at which there's a changepoint and the corresponding
values of the incidence fragility index
}
\description{
Loop over bin.fi for relevant grid of min.p values
}
