% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{bin.fi}
\alias{bin.fi}
\title{Calculates a fragility index for 2x2 data}
\usage{
bin.fi(
  crosstab = NULL,
  X = NULL,
  Y = NULL,
  alg = "greedy",
  test = "fisher",
  alpha = 0.05,
  q = 0,
  r = 0,
  verbose = FALSE,
  fi.start = NULL,
  delta = NULL,
  nsim = 10
)
}
\arguments{
\item{crosstab}{a 2x2 contingency table, stored as a matrix or table. Either input a data table or both X and Y.}

\item{X}{a dataframe representing the covariates. Either input a data table or both X and Y.}

\item{Y}{a dataframe representing the responses. Either input a data table or both X and Y.}

\item{alg}{a string specifying the FI algorithm, 'exact', 'original', 'original.bothdir', 'greedy', or 'hybrid'.
The exact algorithm is described in the \link{Incidence fragility index} article. It will return the exact fragility
index possibly at the cost of longer run time. The original algorithm runs the algorithm proposed by Walsh et al. (2014).
The 'original.bothdir' algorithm runs a natural variant of the original algorithm which allows for searching in both directions
of the chosen group. The greedy algorithm is described in \link{Generalized fragility index} and efficiently upper bounds
the fragility index. The hybrid approach first runs the greedy algorithm and then finetunes the output similar to the exact algorithm.}

\item{test}{a string specifying the test, defaulting to 'fisher' for the Fisher exact test.
Some alternatives are 'fisher.midp', 'pearson.chisq', 'ni.normal'.}

\item{alpha}{a number for the size of test, default 0.05.}

\item{q}{the minimum probability of permitted outcome modifications, defaults to 0 (permitting all modifications).
Note that alg='original' and alg='original.bothdir' do not support this option.}

\item{r}{the minimum probability that a patient combination can reverse significance, default to 0 (reducing to the standard
fragility index). Use alg='greedy' if r > 0.}

\item{verbose}{a logical value for whether to print status updates while running}

\item{fi.start}{the starting fragility index considered in the exact algorithm.}

\item{delta}{the noninferiority margin for when test='ni.normal'}

\item{nsim}{The number of simulations in the root finding algorithm when r>0, by default 10}
}
\value{
a list containing the fragility index and other values, depending on which algorithm is specified.
The name of the fragility index in the list is 'FI'
}
\description{
This is a function which is a wrapper around several general purpose functions to calculate fragility
indices. Several algorithms and several methodologies can be employed to find various fragility indices
for 2x2 tables. The function supports allowing only sufficiently likely modifications, as described in
the article \link{Incidence fragility index}.
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, c(5, 100, 30, 70))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("treatment", "control")
bin.fi(crosstab = x, alg = "greedy", q = .5, verbose = FALSE)$FI
bin.fi(crosstab = x, alg = "exact", q = .01, verbose = FALSE)$FI

}
