% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{draw.binom}
\alias{draw.binom}
\title{Simulate independent draws from two independent Binomial models}
\usage{
draw.binom(ss, theta1 = 0.3, theta2 = 0.7, row.prop = 1/2, matrix = FALSE)
}
\arguments{
\item{ss}{a positive integer representing the sample size}

\item{theta1}{the probability of success of the first group}

\item{theta2}{the probability of success of the second group}

\item{row.prop}{the proportion of observations in the first group. Default is 1/2.}

\item{matrix}{a logical for whether to return a matrix or data frames. Default is FALSE.}
}
\value{
If the argument matrix is FALSE, a list containing a (0,1) group indicator data frame and a (0,1) response data frame,
each with ss rows and one column. If the argument matrix is TRUE, a 2x2 matrix storing the data in a contingency table.
}
\description{
This function can be conveniently used in simulations such as a monte carlo
sample size calculations.
}
\examples{
out <- draw.binom(100)

}
