% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{general.fi.samplesize}
\alias{general.fi.samplesize}
\title{Sample size calculator, taking into account power and fragility index}
\usage{
general.fi.samplesize(
  min.fi = 10,
  min.power = 0.8,
  sample_size_init_power = 100L,
  sample_size_init_fi = NULL,
  get.p.val,
  get.replacements,
  get.sample,
  gamma = 0.6,
  niters = 50,
  cl = NULL,
  verbose = FALSE,
  alpha = 0.05,
  tau = 1/2,
  nsim = 30,
  eps = 0.1,
  algorithm = "walsh"
)
}
\arguments{
\item{min.fi}{the smallest acceptable QUANTILE fragility index. When NULL, the FI calculation is skipped
and sample_size_init_fi is taken to produce the desired FI.}

\item{min.power}{the smallest acceptable power. When NULL, the power calculation is skipped and
sample_size_init_power is taken to produce the desired power.}

\item{sample_size_init_power}{a sample size to initialize the algorithm
(not necessary, defaults to 10) to find the sample size for power}

\item{sample_size_init_fi}{a sample size to initialize the algorithm
(not necessary, defaults to the sample size for power) to find the sample size for fi}

\item{get.p.val}{a function that inputs X and Y and returns a p value}

\item{get.sample}{a function which inputs a sample size and outputs a sample, with
the covariates in a dataframe X and the response in a dataframe Y}

\item{gamma}{the power of n^{-1} in the gradient descent in the Polyak-Ruppert averaging.
The default is 0.60.}

\item{niters}{The number of times to repeat the sample size calculations. The median of the
repitions is then reported.}

\item{cl}{a cluster from the \code{parallel} package, used for the loop over 1:niters}

\item{verbose}{boolean to indicate whether to print out algorithmic information for
the sample size calculation}

\item{alpha}{a number for the size of test}

\item{tau}{the quantile of FI to bound, default 1/2}

\item{nsim}{the number of simulated draws to consider when estimating the power
and expected fragility index quantile}

\item{eps}{a parameter to control the error. The smaller is it, the more precise the
output but the longer the function will take to run.}

\item{algorithm}{A string specifying the algorithm to use to calculate fragility indices.
The default is "greedy". Alternatives include "walsh" for the algorithm originally used by Walsh et al. (2014).}
}
\value{
the length two numeric vector of the calculated sample sizes for the desired power and fragility index
}
\description{
This function takes in a function to compute a p value and a function
to simulate data. Using these, it finds the smallest sample size which
produces a power and fragility index larger than the input
thresholds.
}
\examples{
ss <- general.fi.sscurve(min.fi = 10, min.power = .8, alpha = .05, tau=.5,
get.p.val = function(m)fisher.test(m)$p.value, get.sample = function(ss) draw.binom(ss, matrix=TRUE),
nsim = 10, niters = 2, verbose = FALSE)

}
