% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{general.fi.sscurve}
\alias{general.fi.sscurve}
\title{Fragility index aware sample size calculator for a grid of sample sizes}
\usage{
general.fi.sscurve(
  min.fi,
  get.p.val,
  get.replacements,
  get.sample,
  cl = NULL,
  min.power = 0.8,
  alpha = 0.05,
  verbose = FALSE,
  niters = 5,
  sample_size_init_power = 10L,
  sample_size_init_fi = NULL,
  nsim = 30,
  eps = 0.1,
  tau = 1/2,
  algorithm = "walsh",
  gamma = 0.6
)
}
\arguments{
\item{min.fi}{a vector of the smallest acceptable quantile fragility index. Each element will be passed into}

\item{get.p.val}{a function that inputs X and Y and returns a p value when algorithm is 'greedy' or that
inputs a matrix when algorithm is 'walsh'.}

\item{get.replacements}{a function which outputs a data frame containing all possible row replacements
for Y which are to be considered. The functions inputs the row of Y under consideration,
the row of X under consideration, the row name, and the full original data frames X and Y.
This gets used in the greedy algorithm.}

\item{get.sample}{a function which inputs a sample size and outputs a sample, with
the covariates in a dataframe and the response in a dataframe}

\item{cl}{a cluster from the \code{parallel} package, used in \code{greedy.fi}}

\item{min.power}{the smallest acceptable power}

\item{alpha}{a numeric for the significance cutoff}

\item{verbose}{boolean to indicate whether to print out algorithmic information for the sample size calculation}

\item{niters}{The number of times to repeat the sample size calculations. The median of the
repitions is then reported.}

\item{sample_size_init_power}{a sample size to initialize the algorithm
(not necessary, defaults to 10) to find the sample size for power}

\item{sample_size_init_fi}{a sample size to initialize the algorithm
(not necessary, defaults to the sample size for power) to find the sample size for fi}

\item{nsim}{the number of simulated draws to consider when estimating the power
and expected fragility index}

\item{eps}{a parameter to control the error. The smaller is it, the more precise the
output but the longer the function will take to run.}

\item{tau}{the FI quantile, by default 1/2}

\item{algorithm}{A string, equal to 'walsh' to perform fragility index calculation using the original algorithm
due to Walsh et al. (2014) or equal to 'greedy' to perform fragility index calculation using the greedy algorithm
described in \link{Generalized fragility index}.}

\item{gamma}{the power of n^{-1} in the gradient descent in the Polyak-Ruppert averaging.
The default is 0.60.}
}
\value{
a matrix with 3 columns and row length equal to the length of min.fi. Each row contains the
min.fi value and the output of general.fi.samplesize.
}
\description{
Loops over general.fi.samplesize to calculate the fragility index based sample size.
The p value based sample size will only be calculated once.
}
\examples{
out <- general.fi.sscurve(min.fi = seq(0, 10, by = 5), get.p.val = function(m) fisher.test(m)$p.value,
get.replacements = get.replacements, get.sample = function(ss) draw.binom(ss, matrix=TRUE), niters = 1)

}
