% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{glm.fi}
\alias{glm.fi}
\title{Calculate a fragility index for the coefficient tests in a GLM}
\usage{
glm.fi(
  formula,
  family,
  data,
  max.step = 1,
  alpha = 0.05,
  cl = NULL,
  verbose = FALSE
)
}
\arguments{
\item{formula}{a formula, see the documention of 'glm' for more details}

\item{family}{a description of the error distribution, see the documention of 'glm' for more details.
Currently only gaussian() and binomial() are supported.}

\item{data}{a data frame with measurements on the columns and cases on the rows}

\item{max.step}{an atomic vector for the max step of each response, for when the response type
is restricted}

\item{alpha}{a number for the size of test}

\item{cl}{a cluster from the \code{parallel} package, used in \code{greedy.fi}}

\item{verbose}{a logical value for whether to print status updates while running}
}
\value{
the coefficient table from 'glm' with an extra column containing the fragility index of each coefficient
}
\description{
This is a function which outputs a GLM coefficient table with an extra column
which shows the fragility index corresponding to each p-value. Please note that the
outcome modifications in the gaussian case are determined to be symmetric around the
observation with radius max.step, which is distinct from the sufficiently likely approach
considered throughout the package.
}
\examples{
data(PimaIndiansDiabetes2, package = "mlbench")
primdat <- PimaIndiansDiabetes2[complete.cases(PimaIndiansDiabetes2), ][1:100, ]
glm.fi(diabetes ~ ., binomial(), primdat, verbose = FALSE)

}
