% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{incidence.plot}
\alias{incidence.plot}
\title{Plot the output of bin.fi.incidence}
\usage{
incidence.plot(out)
}
\arguments{
\item{out}{the output of bin.fi.incidence}
}
\value{
a plot visualizing out
}
\description{
Plot the output of bin.fi.incidence
}
\examples{
x <- matrix(nrow=2,byrow=TRUE,c(5, 100, 30, 70),
dimnames = list(c('treatment', 'control'), c('event', 'nonevent')))
out <- bin.fi.incidence(data.table=x, alg='exact')
incidence.plot(out)
}
