% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{ltfu_fi}
\alias{ltfu_fi}
\title{LTFU-aware fragility index calculator}
\usage{
ltfu_fi(
  mat,
  ltfu,
  get.p,
  q = 0,
  alpha = 0.05,
  sampling.control = NULL,
  xlab = "Control group LTFU event rate",
  ylab = "Treatment group LTFU event rate",
  fig.size = 1.1,
  gradientn.scale = 0.99
)
}
\arguments{
\item{mat}{a 2x2 real non-negative integer matrix with control/treatment on the rows and
event/nonevent on the columns.}

\item{ltfu}{a length 2 integer vector counting the number of lost patients in both groups.
The first entry corresponds to the first row of mat.}

\item{get.p}{a function which inputs a 2x2 matrix and returns a p value}

\item{q}{a numeric vector representing the sufficiently likely threshold. A LTFU-aware fragility
index will be calculated for each.}

\item{alpha}{a number for the significance threshold. The default is 0.05}

\item{sampling.control}{a list providing the ndraws and mult arguments for sampling. It is NULL by default.
When NULL, p_o = p_l is taken by default and the distribution of X_l | X_o is available exactly. ndraws is the
number of posterior draws, and mult is the multiplier s (see the LTFU-aware paper).}

\item{xlab}{a string for the horizontal axis label of the output plot}

\item{ylab}{a string for the vertical axis label of the output plot}

\item{fig.size}{A number, by default 1.1. This gives the size of the tiles in the plot. It may need to be
manually tuned.}

\item{gradientn.scale}{A number, by default 0.99. It determines where the tick marks will be for the posterior
probability scale. It may need to be manually tuned.}
}
\value{
Returns a list with two entries. The first \code{FI}, shows the LTFU-aware fragility index
for each supplied entry of q, The last row shows the largest possible q and the corresponding
LTFU-aware fragility index. The second `info``, shows the outcome modifications that make each
LTFU-aware fragility index possible.
}
\description{
LTFU-aware fragility index calculator
}
\examples{
mat <- matrix(nrow=2, byrow=TRUE, c(315, (1198+314)-315, 350, (1189+328)-350))
ltfu <- c(24+3, 27+5)
out <- ltfu_fi(mat, ltfu, function(m) fisher.test(m)$p.value)
}
