% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{min.fi.curve}
\alias{min.fi.curve}
\title{Loops over general.fi.samplesize for a grid of sample sizes.}
\usage{
\method{min}{fi.curve}(
  min.fi,
  get.p.val,
  get.replacements,
  get.sample,
  cl = NULL,
  min.power = 0.8,
  alpha = 0.05,
  verbose = FALSE,
  niters = 5,
  sample_size_init_power = 10L,
  sample_size_init_fi = NULL,
  nsim = 30,
  eps = 0.1,
  tau = 1/2,
  algorithm = "greedy",
  gamma = 0.6
)
}
\arguments{
\item{min.fi}{a vector of the smallest acceptable quantile fragility index}

\item{get.p.val}{a function that inputs X and Y and returns a p value}

\item{get.replacements}{a function that inputs a response, covariate, and row name and outputs the
replacements to be considered in the algorithm}

\item{get.sample}{a function which inputs a sample size and outputs a sample, with
the covariates in a dataframe and the response in a dataframe}

\item{cl}{a cluster from the \code{parallel} package, used in \code{greedy.fi}}

\item{min.power}{the smallest acceptable power}

\item{alpha}{a number for the size of test}

\item{verbose}{boolean to indicate whether to print out algorithmic information for the sample size calculation}

\item{niters}{The number of times to repeat the sample size calculations. The median of the
repitions is then reported.}

\item{sample_size_init_power}{a sample size to initialize the algorithm
(not necessary, defaults to 10) to find the sample size for power}

\item{sample_size_init_fi}{a sample size to initialize the algorithm
(not necessary, defaults to the sample size for power) to find the sample size for fi}

\item{nsim}{the number of simulated draws to consider when estimating the power
and expected fragility index}

\item{eps}{a parameter to control the error. The smaller is it, the more precise the
output but the longer the function will take to run.}

\item{tau}{the FI quantile, by default 1/2}

\item{algorithm}{whether to short circuit the call to greedy.fi and instead use
a binary clasic alg. Default is 'greedy'}

\item{gamma}{the power of n^{-1} in the gradient descent in the Polyak-Ruppert averaging.
The default is 0.60.}
}
\value{
the calculated sample size for the desired power and fragility index
}
\description{
Loops over general.fi.samplesize for a grid of sample sizes.
}
\examples{
get.p.val <- function(X, Y) {
  x <- X[[1]]
  y <- Y[[1]]
  mat <- matrix(nrow = 2, ncol = 2)
  mat[1, 1] <- sum((x == 0) & (y == 0))
  mat[1, 2] <- sum((x == 0) & (y == 1))
  mat[2, 1] <- sum((x == 1) & (y == 0))
  mat[2, 2] <- sum((x == 1) & (y == 1))
  fisher.test(mat)$p.value
}
get.replacements <- list(function(y, x, rn, Y, X) setdiff(unique(Y[[1]]), y))
get.sample <- draw.binom

out <- min.fi.curve(
  min.fi = seq(0, 10, by = 5), get.p.val = get.p.val, get.replacements = get.replacements,
  get.sample = get.sample, niters = 1, algorithm = "greedy"
)
}
