% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{normal.sl}
\alias{normal.sl}
\title{Calculate level of smallest HDR which contains the modified observation}
\usage{
normal.sl(y, y.mod)
}
\arguments{
\item{y}{a numeric vector of outcomes}

\item{y.mod}{a numeric vector of outcomes after the fragility index modification}
}
\value{
a scalar numeric: a ratio of the likilihoods in y and y.mod, estimated
using the MLE with the data y
}
\description{
Used for internal functions in the study of generalized fragility indices.
}
\examples{
y <- rnorm(100, 0, 1)
y.mod <- rnorm(100,.1, 1.1) # should be returned from an FI procedure
normal.sl(y, y.mod)

}
