% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{rir}
\alias{rir}
\title{Bayesian RIR calculation}
\usage{
rir(
  mat,
  get.p = function(mat) fisher.test(mat)$p.value,
  alpha = 0.05,
  iters = 10000,
  cl = NULL
)
}
\arguments{
\item{mat}{a 2 x 2 contingency table of counts. The treatment should be on the first row and the event should be on
the first column}

\item{get.p}{a function which inputs a 2x2 matrix and outputs a p value (numeric)}

\item{alpha}{a numeric for the significance threshold, by default 0.05}

\item{iters}{the number of monte carlo iterations, by default 10000}

\item{cl}{A parallel cluster to parellelize the monte carlo iterates, by default NULL}
}
\value{
a length 2 numeric vector, giving the conditional posterior mean number of replacements and
the posterior probability of reversing not being possible
}
\description{
The Bayesian RIR is a Bayesian view of the the Robustness of an Inference to Replacement proposed by Frank et al. (2021).
Unlike the RIR, it takes the posterior expectation of the final quantity in the definition of the RIR. This is explained
in more detail in a letter to the editor to Frank et al. (2021) in the Journal of Clinical Epidemiology.
}
\examples{
walter.mat <- structure(c(0, 5, 96, 90), .Dim = c(2L, 2L)) # data from Walter, in correct formatting
rir(walter.mat)

}
