% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{surv.fi}
\alias{surv.fi}
\title{Calculate a fragility index for data with right-censored survival response}
\usage{
surv.fi(
  time,
  status,
  group,
  test = "logrank",
  max.step = NULL,
  cl = NULL,
  alpha = 0.05,
  tau = NULL,
  verbose = FALSE
)
}
\arguments{
\item{time}{the time of either an event or right censoring}

\item{status}{a 0,1 variable, with 1 if event and 0 if right-censored}

\item{group}{a factor with levels representing group membership}

\item{test}{a string specifying the test type, default is 'logrank'}

\item{max.step}{a numeric for the maximum allowable change of the time outcome for each
patient; note that times will be thresholded to 0 if negative}

\item{cl}{a cluster from the \code{parallel} package, used to compute fragility index over
each modified observation at each stage of the greedy algorithm}

\item{alpha}{a number for the size of test}

\item{tau}{an optional parameter for the rmst difference test}

\item{verbose}{a logical value for whether to print status updates while running}
}
\value{
the output of greedy.fi for the given test
}
\description{
This is a function which is a wrapper around internal functions which calculate
the fragility index for different data types and test specifications. It is used
for survival data.
}
\examples{
dat <- get.survival.data(100, 6)
cl <- parallel::makeCluster(parallel::detectCores()-2)
out <- surv.fi(dat$time, dat$status, dat$group, cl=cl)
parallel::stopCluster(cl)
}
