% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bin.fi.exact}
\alias{bin.fi.exact}
\title{Exactly calculate a fragility index for 2x2 data (deprecated)}
\usage{
bin.fi.exact(crosstab, get.p, alpha = 0.05, max.f = Inf, verbose = FALSE)
}
\arguments{
\item{crosstab}{a 2x2 contingency table with treatments on rows and outcomes on columns}

\item{get.p}{a function inputting a table and outputting a p-value}

\item{alpha}{a string whose value being "exact" leads to an exact calculation}

\item{max.f}{the maximum fragility index considered}

\item{verbose}{a boolean value for whether to print progress
percentages (in increments of roughly ten percent)}
}
\value{
a list containing the signed fragility index and other accompanying values,
similar to \code{greedy.fi}.
}
\description{
Exactly calculate a fragility index using a 2x2 table.
Do a grid search over all FI values <= some value. This has
been deprecated and will stop with an error to use
bin.fi.exact2 instead, as it is more efficient.
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, rgeom(4, 1 / 50))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("control", "treatment")
get.p <- function(tab) fisher.test(tab)$p.value
FragilityTools:::bin.fi.exact(x, get.p)
}
