% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bin.fi.exact2}
\alias{bin.fi.exact2}
\title{Exactly calculate a fragility index for 2x2 data (2.0)}
\usage{
bin.fi.exact2(
  crosstab,
  get.p,
  alpha = 0.05,
  fi.start = 1,
  can.vary = matrix(rep(TRUE, 4), nrow = 2),
  start.reversed = FALSE,
  start.p = NA
)
}
\arguments{
\item{crosstab}{a 2x2 contingency table with interventions on rows and outcomes on columns}

\item{get.p}{a function inputting a table and outputting a p-value}

\item{alpha}{a scalar numeric for the significance cutoff}

\item{fi.start}{The starting fi considered, by default 1.}

\item{can.vary}{a 2x2 boolean matrix for whether each entry can decrease. See the
\link{incidence fragility} article for explanation.}

\item{start.reversed}{a boolean for whether the start.fi value already reverses
statistical significance. Note, FALSE just means unknown. (note to self: i should
update this to be true/false/na)}

\item{start.p}{The p value of the table with reversed statistical significance, paired with
the parameter start.reversed}
}
\value{
a list containing the signed fragility index and other accompanying values,
similar to \code{greedy.fi}
}
\description{
Do not call this function: use bin.fi with option 'exact' instead
The algorithm starts at a given FI value and radiates outward from there.
Note this can use an increasing or decreasing search, depending on the input.
}
\details{
Do not call this function: use bin.fi with option 'exact' instead. If bin.fi.exact2 is
initialized with too large of a value, the algorithm could malfunction. It's best to
leave the default behavior unless sure.
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, rgeom(4, 1 / 50))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("control", "treatment")
get.p <- function(tab) fisher.test(tab)$p.value
bin.fi.exact2(x, get.p)
}
