% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bin.fi.greedy}
\alias{bin.fi.greedy}
\title{Greedy calculation of fragility indices for 2x2 contingency tables}
\usage{
bin.fi.greedy(
  crosstab,
  get.p,
  alpha = 0.05,
  can.vary = matrix(rep(TRUE, 4), nrow = 2)
)
}
\arguments{
\item{crosstab}{a 2x2 contingency table with interventions on rows and outcomes on columns}

\item{get.p}{a function inputting a table and outputting a p-value}

\item{alpha}{a scalar numeric for the significance cutoff, by default 0.05}

\item{can.vary}{a 2x2 boolean matrix for whether each entry can decrease. See the
\link{incidence fragility index} article for explanation.}
}
\value{
a list containing the signed fragility index, in addition to the modified contingency
table which has reversed statistical significance and the p value sequence (as in \code{greedy.fi})
}
\description{
This is an internal function, please use bin.fi instead
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, rgeom(4, 1 / 50))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("control", "treatment")
get.p <- function(tab) fisher.test(tab)$p.value
FragilityTools:::bin.fi.greedy(x, get.p)
}
