% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{bin.fi.incidence}
\alias{bin.fi.incidence}
\title{Calculate all of the incidence fragility indices}
\usage{
bin.fi.incidence(
  crosstab = NULL,
  alg = "exact",
  test = "fisher",
  fi.start = NULL,
  alpha = 0.05,
  verbose = FALSE,
  delta = NULL,
  X = NULL,
  Y = NULL
)
}
\arguments{
\item{crosstab}{a 2x2 non negative integer matrix (contingency table) representing the output of a clinical trial.
The rows correspond to treatment/control and the columns to event/nonevent.}

\item{alg}{a string specifying the FI algorithm, 'exact', 'greedy', 'hybrid'.
See the documentation of bin.fi for more information. Note that 'original' and 'original.bothdir'
do not support the per-patient sufficiently likely threshold so cannot be used here.}

\item{test}{a string specifying the test, defaulting to 'fisher' for the Fisher exact test.
Alternatives include 'fisher.midp' and 'pearson.chisq'.}

\item{fi.start}{the starting fragility index considered in the exact algorithm}

\item{alpha}{a numeric for the significance threshold of the statistical test}

\item{verbose}{a logical value for whether to print status updates while running}
}
\value{
a matrix containing the sufficiently likely threshold values q at which there is
a changepoint of the corresponding values of the incidence fragility index
}
\description{
The function is used to calculate the incidence fragility indices for every possible sufficiently likely threshold.
Note, the incidence fragility indices are formally defined in the article \link{Incidence fragility index}. The function bin.fi
in this package will calculate an incidence fragility index for a given threshold, and this function loops over bin.fi
for relevant grid of sufficiently likely thresholds min.p (i.e. q). The function \code{incidence.plot} allows for a convenient
visualization of these values.
}
\examples{
x <- matrix(nrow=2,byrow=TRUE,c(5, 100, 30, 70),
dimnames = list(c('treatment', 'control'), c('event', 'nonevent')))
out <- bin.fi.incidence(crosstab=x, alg='exact')

}
