% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{bin.fi}
\alias{bin.fi}
\title{Calculate a fragility index for data with binary responses}
\usage{
bin.fi(
  data.table = NULL,
  X = NULL,
  Y = NULL,
  alg = "greedy",
  test = "fisher",
  fi.start = NULL,
  alpha = 0.05,
  verbose = FALSE,
  warm.start = TRUE,
  delta = NULL,
  min.p = 0
)
}
\arguments{
\item{data.table}{a 2x2 contingency table, stored as a matrix or table}

\item{X}{a dataframe representing the covariates}

\item{Y}{a dataframe representing the responses}

\item{alg}{a string specifying the FI algorithm, 'greedy' (default), 'exact', 'original', or 'original.bothdir'}

\item{test}{a string specifying the test, defaulting to 'fisher' for the Fisher exact test.
An alternative is 'fisher.midp' and 'pearson.chisq'.}

\item{fi.start}{the starting fragility index considered in the exact algorithm.}

\item{alpha}{a number for the size of test}

\item{verbose}{a logical value for whether to print status updates while running}

\item{warm.start}{If using the exact algorithm without specifying a \code{max.f} argument,
a greedy algorithm will be ran first to upper bound the search for the exact algorithm.
Set to FALSE to only run the exact algorithm.}

\item{delta}{the noninferiority margin for when test='ni.normal'}

\item{min.p}{the minimum probability of allowable outcome changes, defaults to 0 (allowing all changes).
Note that alg='original' or alg='original.bothdir' does not support this option.}
}
\value{
the fragility index and other values, depending on which algorithm is specified.
}
\description{
Estimate a fragility index using for the fisher exact test using a table
}
\details{
This is a function which is a wrapper around internal functions which calculate
the fragility index for different data types and test specifications.
}
\examples{
x <- matrix(nrow = 2, byrow = TRUE, c(5, 100, 30, 70))
colnames(x) <- c("event", "nonevent")
rownames(x) <- c("treatment", "control")
p.grid <- seq(0, 1, by = .01)
fi.grid <- sapply(p.grid, function(p) bin.fi(data.table = x, alg = "greedy", min.p = p, verbose = FALSE)$FI)
ggplot2::qplot(p.grid[fi.grid < Inf], fi.grid[fi.grid < Inf],
  xlab = "Likelihood bound",
  ylab = "Fragility index", xlim = c(0, 1), main = "Fisher exact test fragility indices (n=205)"
)
}
