% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{bin.fi}
\alias{bin.fi}
\title{Calculate a fragility index for data with binary responses}
\usage{
bin.fi(
  data.table = NULL,
  X = NULL,
  Y = NULL,
  alg = "greedy",
  test = "fisher",
  max.f = Inf,
  alpha = 0.05,
  verbose = FALSE,
  warm.start = TRUE,
  delta = NULL
)
}
\arguments{
\item{data.table}{a 2x2 contingency table, stored as a matrix or table}

\item{X}{a dataframe representing the covariates}

\item{Y}{a dataframe representing the responses}

\item{alg}{a string specifying the FI algorithm, 'greedy' (default), 'exact', or 'walsh'}

\item{test}{a string specifying the test, defaulting to 'fisher' for the Fisher exact test.
An alternative is 'fisher.midp' and 'pearson.chisq'.}

\item{max.f}{the maximum fragility index considered}

\item{alpha}{a number for the size of test}

\item{verbose}{a logical value for whether to print status updates while running}

\item{warm.start}{If using the exact algorithm without specifying a \code{max.f} argument,
a greedy algorithm will be ran first to upper bound the search for the exact algorithm.
Set to FALSE to only run the exact algorithm.}

\item{delta}{the noninferiority margin for when test='ni.normal'}
}
\value{
the output of greedy.fi for the given test
}
\description{
Estimate a fragility index using for the fisher exact test using a table
}
\details{
This is a function which is a wrapper around internal functions which calculate
the fragility index for different data types and test specifications.
}
\examples{
x <- matrix(nrow=2,byrow=TRUE,rgeom(4, 1/50))
colnames(x) <- c('event', 'nonevent')
rownames(x) <- c('control', 'treatment')
out <- bin.fi(data.table=x, alg='exact')

}
