% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bin.fi.walsh}
\alias{bin.fi.walsh}
\title{Original FI, for binary data}
\usage{
bin.fi.walsh(mat, get.p, alpha = 0.05)
}
\arguments{
\item{mat}{a 2x2 matrix with groups on the rows}

\item{get.p}{a function which accepts a 2x2 matrix and outputs a p value}

\item{alpha}{the p-value cutoff to reject, 0.05 by default}
}
\description{
Original FI, for binary data
}
\examples{
bin.fi.walsh(matrix(c(100,96,13,28),nrow=2), function(mat)fisher.test(mat)$p.value)

}
