% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{binmeta.fi}
\alias{binmeta.fi}
\title{Calculate a fragility index for a meta analysis with binary response}
\usage{
binmeta.fi(
  data,
  restr.study = NULL,
  method = "MH",
  sm = "RR",
  cl = NULL,
  alpha = 0.05,
  verbose = FALSE,
  q = 0
)
}
\arguments{
\item{data}{a dataframe with the number of events and observations for the control and treatment
groups as columns. The names of the studies must be the rownames, and the names cannot contain two
consuctive dots. The order should be event.e, n.e, event.c, n.c, as in meta::binmeta}

\item{restr.study}{a character vector giving the names of the studies whose patients are allowed
to have response changes. The default is all the studies.}

\item{method}{A character string indicating which method is to be used for pooling of studies.
One of "Inverse", "MH", "Peto", "GLMM", or "SSW", can be abbreviated. (From \code{meta}.`)}

\item{sm}{A character string indicating which summary measure ("RR", "OR", "RD", or "ASD").
See the `meta`` package documention for more details.}

\item{cl}{a cluster from the \code{parallel} package, used to compute fragility index over
each modified observation at each stage of the greedy algorithm}

\item{alpha}{a number for the size of test}

\item{verbose}{a logical value for whether to print status updates while running}

\item{q}{the sufficiently likely threshold, defaults to 0, permitting all modifications. Any
empirical conditional outcome probability strictly lower than q will not be permitted to
receive outcome modifications.}
}
\value{
the output of greedy.fi specialized to the meta analysis case
}
\description{
This is a function which is a wrapper around internal functions which calculate
the fragility index for different data types and test specifications. It is used
for meta analyses on binary data. This function makes use of p-value calculators
from the \code{meta} R package.
}
\examples{
data(Olkin95, package = "meta")
out <- binmeta.fi(data = Olkin95[1:5, -2], restr.study = c("Dewar", "Lippschutz"))
out$FI

}
