% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{get.rejection.rates}
\alias{get.rejection.rates}
\title{Get rejection rates of the two main tests}
\usage{
get.rejection.rates(
  get.p.val,
  get.replacements = NULL,
  get.sample.null = NULL,
  get.sample.alt = NULL,
  phi,
  n,
  alpha = 0.05,
  algorithm = "greedy",
  cl = NULL,
  nsim = 1000
)
}
\arguments{
\item{get.p.val}{a}

\item{get.replacements}{a}

\item{get.sample.null}{a set to NULL to skip null calculations}

\item{get.sample.alt}{a set to NULL to skip alternative calculations}

\item{phi}{a}

\item{n}{a}

\item{alpha}{a}

\item{algorithm}{a}
}
\value{
a length 4 vector with the size and power of both tests.
}
\description{
Get rejection rates of the two main tests
}
\examples{
get.rejection.rates(TODO)

}
