% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.survival.data}
\alias{get.survival.data}
\title{Simulate right-censored survival responses in two groups}
\usage{
get.survival.data(n, num.times, p1 = NULL, p2 = NULL)
}
\arguments{
\item{n}{the number of observations}

\item{num.times}{the number of time points at which events or right-censors can occur}

\item{p1}{the proportion of events in the first group, by default NULL which takes p1 to be random}

\item{p2}{the proportion of events in the second group, by default NULL which takes p1 to be random}
}
\value{
a dataframe with three columns: group, time, and status
}
\description{
Simulate right-censored survival responses in two groups
}
\examples{
dat <- get.survival.data(n = 100, num.times = 6)
}
