% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_beta_parameters}
\alias{get_beta_parameters}
\title{Gets parameters for sampling in ltfu_fi
Chooses them so that the mode is po and the 75th
quantile is mult*po}
\usage{
get_beta_parameters(po, mult = 1.3)
}
\description{
Gets parameters for sampling in ltfu_fi
Chooses them so that the mode is po and the 75th
quantile is mult*po
}
