% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{greedy.fi}
\alias{greedy.fi}
\title{Calculate a fragility index using a greedy algorithm}
\usage{
greedy.fi(
  X,
  Y,
  get.replacements,
  get.p.val,
  alpha = 0.05,
  verbose = FALSE,
  cl = NULL,
  only.consider = c(),
  dont.consider = c()
)
}
\arguments{
\item{X}{a data frame of covariates}

\item{Y}{a data frame of responses}

\item{get.replacements}{a list of of function that inputs a response, covariate, and row name and
outputs the replacements to be considered in the algorithm (one function for each response column).
If there is only one column, a function can be passed in instead of a list containing the one function.}

\item{get.p.val}{a function that inputs X and Y and returns a p value}

\item{alpha}{a number for the size of test}

\item{verbose}{a logical value for whether to print status updates while running}

\item{cl}{a cluster from the \code{parallel} package, used to compute fragility index over
each modified observation at each stage of the greedy algorithm}

\item{only.consider}{a vector of row names to only consider}

\item{dont.consider}{a vector of row names to not consider}
}
\value{
a list containing a fragility index and related information
\describe{
\item{FI}{A fragility index, positive if the initial $p$-value was < alpha and negative otherwise}
\item{p_val_sequence}{An atomic vectors of the p-values at each step of the iteratative algorithm,
where the first is the starting p value and the last is the p value which crossed the alpha
threshold}
\item{reverse}{A boolean indicating whether the original p value is larger than alpha}
\item{num_patients}{The number of patients whose responses were subject to change}
\item{patients}{An atomic vector with the rownames of the rows which had their responses changed}
\item{old_responses}{A list of dataframes of the original responses of each changed patient}
\item{new_responses}{A list of dataframes of the new responses of each changed patient}
}
}
\description{
Estimate a fragility index using covariates, response(s), and a p value function.
}
\details{
This is a general function which uses a greedy algorithm to compute a fragility
index.The function arguments accept data from two sources: covariates in X and
responses in Y. Covariates are unchanged by the algorithm, while responses are
iteratively changed. The type of each response is specified to determine which
substitute outcomes to consider.
}
\examples{
n <- 100
X <- data.frame('tr_group'=sample(c('treated', 'not treated'), n, TRUE))
Y <- data.frame('outcome'=sample(c('sick', 'healthy'), n, TRUE))
get.p.val <- function(X,Y) fisher.test(table(X[[1]],Y[[1]]))$p.value
get.replacements <- list(function(y,x,rn,Y,X) setdiff(unique(Y[[1]]), y))
greedy.fi(X, Y, get.p.val = get.p.val, get.replacements=get.replacements)
}
