% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mat_to_xy}
\alias{mat_to_xy}
\title{Convert a 2x2 table to a X,Y data frames}
\usage{
mat_to_xy(mat, event.col = 1)
}
\arguments{
\item{mat}{2x2 numeric matrix}
}
\value{
list containing two one-column data frames X and Y
}
\description{
Convert a 2x2 table to a X,Y data frames
}
\examples{
mat_to_xy(matrix(1:4, nrow = 2))
}
